+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog][keepachangelog] and this project adheres to [Semantic Versioning][semver].
 
+## [v2.4.0] - 2018-06-27
+
+### Added
+
+- Options array for `->user()->report()->make()` method
+- Options array for `->user()->report()->refresh()` method
+
 ## [v2.3.0] - 2018-06-27
 
 ### Added
@@ -33,6 +40,7 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 - API backward compatible fix with passing empty data (for example - on calling `_refresh` method)
 
+[v2.4.0]:https://github.com/avto-dev/b2b-api-php/compare/v2.3.0...v2.4.0
 [v2.3.0]:https://github.com/avto-dev/b2b-api-php/compare/v2.2.0...v2.3.0
 [v2.2.0]:https://github.com/avto-dev/b2b-api-php/compare/v2.1.5...v2.2.0
 [v2.1.5]:https://github.com/avto-dev/b2b-api-php/compare/v2.1.4...v2.1.5